+++ src/main/java/com/picimako/mockitools/intention/ConvertMockSpyFieldToCallIntention.java
@@ -2,28 +2,31 @@
 
 package com.picimako.mockitools.intention;
 
-import static com.picimako.mockitools.util.ListPopupHelper.selectItemAndRun;
+import static com.intellij.openapi.command.WriteCommandAction.runWriteCommandAction;
+import static com.picimako.mockitools.dsl.MockAnnotation.isAttributeEnabledOnMockAnnotation;
 import static com.picimako.mockitools.MockitoQualifiedNames.ANSWER;
 import static com.picimako.mockitools.MockitoQualifiedNames.DEFAULT_ANSWER;
 import static com.picimako.mockitools.MockitoQualifiedNames.EXTRA_INTERFACES;
 import static com.picimako.mockitools.MockitoQualifiedNames.MOCK;
+import static com.picimako.mockitools.MockitoQualifiedNames.MOCK_MAKER;
 import static com.picimako.mockitools.MockitoQualifiedNames.NAME;
 import static com.picimako.mockitools.MockitoQualifiedNames.ORG_MOCKITO_MOCK;
 import static com.picimako.mockitools.MockitoQualifiedNames.ORG_MOCKITO_MOCKITO;
+import static com.picimako.mockitools.MockitoQualifiedNames.ORG_MOCKITO_QUALITY_STRICTNESS;
 import static com.picimako.mockitools.MockitoQualifiedNames.ORG_MOCKITO_SPY;
 import static com.picimako.mockitools.MockitoQualifiedNames.SPY;
+import static com.picimako.mockitools.MockitoQualifiedNames.STRICTNESS;
 import static com.picimako.mockitools.MockitoolsPsiUtil.MOCKITO_WITH_SETTINGS;
 import static com.picimako.mockitools.intention.ConvertMockCallToFieldIntention.isDefaultAnswer;
 import static com.picimako.mockitools.intention.MethodRearranger.reOrder;
+import static com.picimako.mockitools.util.ListPopupHelper.selectItemAndRun;
 import static java.util.stream.Collectors.joining;
 
-import com.intellij.codeInsight.AnnotationUtil;
 import com.intellij.codeInsight.intention.IntentionAction;
 import com.intellij.codeInspection.util.IntentionFamilyName;
 import com.intellij.codeInspection.util.IntentionName;
 import com.intellij.ide.highlighter.JavaFileType;
 import com.intellij.lang.jvm.annotation.JvmAnnotationAttribute;
-import com.intellij.openapi.command.WriteCommandAction;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.JavaPsiFacade;
@@ -34,6 +37,7 @@ import com.intellij.psi.PsiClass;
 import com.intellij.psi.PsiCodeBlock;
 import com.intellij.psi.PsiDocumentManager;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiEnumConstant;
 import com.intellij.psi.PsiField;
 import com.intellij.psi.PsiFile;
 import com.intellij.psi.PsiIdentifier;
@@ -45,9 +49,9 @@ import com.intellij.psi.PsiReferenceExpression;
 import com.intellij.psi.PsiTypeElement;
 import com.intellij.util.IncorrectOperationException;
 import com.picimako.mockitools.MockitoQualifiedNames;
-import com.picimako.mockitools.util.PsiClassUtil;
 import com.picimako.mockitools.intention.MethodRearranger.ClassMethodCellRenderer;
 import com.picimako.mockitools.resources.MockitoolsBundle;
+import com.picimako.mockitools.util.PsiClassUtil;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Arrays;
@@ -154,7 +158,7 @@ public class ConvertMockSpyFieldToCallIntention implements IntentionAction {
             mockitoMockingCall = new StringBuilder("Mockito").append(".").append(MOCK).append("(");
             appendType(fieldToConvert, mockitoMockingCall);
 
-            PsiAnnotation mockAnnotation = fieldToConvert.getAnnotation(ORG_MOCKITO_MOCK);
+            var mockAnnotation = fieldToConvert.getAnnotation(ORG_MOCKITO_MOCK);
             //Assemble Mockito.mock() call based on the @Mock annotation attributes
             boolean isMockSettingsOverride = true;
             //Handle .mock(Class, Answer) and .mock(Class, String)
@@ -177,10 +181,10 @@ public class ConvertMockSpyFieldToCallIntention implements IntentionAction {
                 StringBuilder mockSettings = new StringBuilder("Mockito").append(".withSettings()");
 
                 //Handle boolean attributes
-                BOOLEAN_ATTRIBUTES.forEach(attributeName ->
-                    Optional.ofNullable(AnnotationUtil.getBooleanAttributeValue(mockAnnotation, attributeName))
-                        .filter(Boolean.TRUE::equals) //Given that the default values are false, and to be true they have to be specified explicitly
-                        .ifPresent(enabled -> appendSetting(mockSettings, attributeName, "")));
+                BOOLEAN_ATTRIBUTES.forEach(attributeName -> {
+                    if (isAttributeEnabledOnMockAnnotation(mockAnnotation, attributeName))
+                        appendSetting(mockSettings, attributeName, "");
+                });
 
                 //Handle attributes for whose MockSettings counterpart the value of the attribute must be passed in.
                 //E.g. withSettings().name("some name") or withSettings().answer(anAnswer)
@@ -207,6 +211,23 @@ public class ConvertMockSpyFieldToCallIntention implements IntentionAction {
                     }
                 });
 
+                valueOf(mockAnnotation.findAttribute(STRICTNESS)).ifPresent(value -> {
+                    if (value instanceof PsiReferenceExpression) {
+                        var resolved = ((PsiReferenceExpression) value).resolve();
+                        if (resolved instanceof PsiEnumConstant) {
+                            String strictnessName = ((PsiEnumConstant) resolved).getName();
+                            //Mock.Strictness.TEST_LEVEL_DEFAULT has no matching enum constant in Strictness, thus we'll ignore it.
+                            if (!"TEST_LEVEL_DEFAULT".equals(strictnessName)) {
+                                runWriteCommandAction(file.getProject(),
+                                    () -> PsiClassUtil.importClass(ORG_MOCKITO_QUALITY_STRICTNESS, mockAnnotation));
+                                appendSetting(mockSettings, STRICTNESS, "Strictness." + strictnessName);
+                            }
+                        }
+                    }
+                });
+
+                valueOf(mockAnnotation.findAttribute(MOCK_MAKER)).ifPresent(value -> appendSetting(mockSettings, MOCK_MAKER, value.getText()));
+
                 mockitoMockingCall.append(", ").append(mockSettings); //Adds the second parameter to Mockito.mock(Class, MockSettings)
             }
         } else { //@Spy -> Mockito.spy()
@@ -220,7 +241,7 @@ public class ConvertMockSpyFieldToCallIntention implements IntentionAction {
 
         mockitoMockingCall.append(")");
 
-        WriteCommandAction.runWriteCommandAction(file.getProject(), () -> {
+        runWriteCommandAction(file.getProject(), () -> {
             var elementFactory = JavaPsiFacade.getElementFactory(file.getProject());
             PsiClassUtil.importClass(ORG_MOCKITO_MOCKITO, file);
             var mockitoMockingInitializer = elementFactory.createExpressionFromText(mockitoMockingCall.toString(), file);
@@ -238,11 +259,10 @@ public class ConvertMockSpyFieldToCallIntention implements IntentionAction {
             var mockitoMockingVariableDeclaration = elementFactory.createVariableDeclarationStatement(fieldToConvert.getName(), fieldToConvert.getType(), finalMockingInitializer);
             PsiCodeBlock methodBody = targetMethod.getBody();
             if (methodBody != null) {
-                if (methodBody.getFirstBodyElement() != null) {
+                if (methodBody.getFirstBodyElement() != null)
                     methodBody.addBefore(mockitoMockingVariableDeclaration, methodBody.getFirstBodyElement());
-                } else {
+                else
                     methodBody.add(mockitoMockingVariableDeclaration);
-                }
             }
             fieldToConvert.delete();
         });
@@ -262,8 +282,8 @@ public class ConvertMockSpyFieldToCallIntention implements IntentionAction {
             : Optional.empty();
     }
 
-    private void appendSetting(StringBuilder sb, String attributeName, String argument) {
-        sb.append(".").append(attributeName).append("(").append(argument).append(")");
+    private void appendSetting(StringBuilder sb, String methodName, String argument) {
+        sb.append(".").append(methodName).append("(").append(argument).append(")");
     }
 
     /**